move Whisper smoke tests to test_cli_whisper_smoke.py
